+++ tatk/policy/vec.py
