+++ tatk/policy/vector/vector_multiwoz.py
@@ -2,7 +2,7 @@
 import os
 import json
 import numpy as np
-from tatk.policy.vector import Vector
+from tatk.policy.vec import Vector
 from tatk.util.multiwoz.lexicalize import delexicalize_da, flat_da, deflat_da, lexicalize_da
 from tatk.util.multiwoz.multiwoz_slot_trans import REF_USR_DA
 from tatk.util.multiwoz.dbquery import query
